GOVSI-1114: Refactor common Notify test code into base class
- Create a new abstract test class that contains the code for setting up a Notify stub, generating messages and a test configuration
- Allow Notify stub to choose a random port rather than fixing it to `8888`, this should aid portability and stop test classes contaminating each other
